refactor: Room 엔티티 리팩토링
